Watts 9: Delete user deletes all user stuff in db

+++ WattsApp/app/src/main/java/com/dabloons/wattsapp/manager/UserManager.java
@@ -3,15 +3,18 @@ package com.dabloons.wattsapp.manager;
 import android.content.Context;
 import android.util.Log;
 
+import com.dabloons.wattsapp.WattsApplication;
 import com.dabloons.wattsapp.model.integration.IntegrationAuth;
 import com.dabloons.wattsapp.model.integration.IntegrationType;
 import com.dabloons.wattsapp.model.User;
 import com.dabloons.wattsapp.model.integration.NanoleafPanelAuthCollection;
 import com.dabloons.wattsapp.model.integration.NanoleafPanelIntegrationAuth;
 import com.dabloons.wattsapp.model.integration.PhillipsHueIntegrationAuth;
+import com.dabloons.wattsapp.repository.LightRepository;
 import com.dabloons.wattsapp.repository.UserAuthRepository;
 import com.dabloons.wattsapp.repository.UserRepository;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
 import java.util.List;
@@ -28,6 +31,7 @@ public class UserManager {
     private UserRepository userRepository;
     private UserAuthRepository userAuthRepository;
 
+    // Todo: Make every singleton class like this (private constructor, vars initalized in constructor)
     private UserManager() {
         userRepository = UserRepository.getInstance();
         userAuthRepository = UserAuthRepository.getInstance();
@@ -104,12 +108,40 @@ public class UserManager {
         userAuthRepository.getUserIntegrations(callback);
     }
 
-    public Task<Void> deleteUser(Context context){
+    public void deleteUser(Context context, WattsCallback<Void, Void> callback){
         // Delete the user account from the Firestore
-        String uid = this.getCurrentUser().getUid();
-        return userRepository.deleteUserFromFirestore().addOnCompleteListener(task -> {
-            // Once done, delete the user data from Auth
-            userRepository.deleteUser(context);
+        this.deleteUserEntities((var, status) -> {
+            if(!status.success) {
+                Log.e(LOG_TAG, status.message);
+                callback.apply(null, new WattsCallbackStatus(false, status.message));
+                return null;
+            }
+
+            userAuthRepository.deleteIntegrationsForUser((var1, status1) -> {
+                if(!status.success) {
+                    Log.e(LOG_TAG, status.message);
+                    callback.apply(null, new WattsCallbackStatus(false, status.message));
+                    return null;
+                }
+
+                userRepository.deleteUserFromFirestore()
+                        .addOnCompleteListener(task -> {
+                            userRepository.deleteUser(context)
+                                    .addOnCompleteListener(task1 -> {
+                                        userRepository.signOut(WattsApplication.getAppContext());
+                                        callback.apply(null, new WattsCallbackStatus(true));
+                                    })
+                                    .addOnFailureListener(task1 -> {
+                                        callback.apply(null, new WattsCallbackStatus(false, task1.getMessage()));
+                                    });
+                        })
+                        .addOnFailureListener(task -> {
+                            callback.apply(null, new WattsCallbackStatus(false, task.getMessage()));
+                        });
+                return null;
+            });
+
+            return null;
         });
     }
 
@@ -127,6 +159,39 @@ public class UserManager {
 
     public Task<Void> setAuthPropString(String prop, String value, IntegrationType type) {
         return userAuthRepository.updatePropertyString(prop, value, type);
+    }
+
+    private void deleteUserEntities(WattsCallback<Void, Void> callback) {
+        RoomManager.getInstance().deleteRoomsForUser((var, status) -> {
+            if(!status.success) {
+                Log.e(LOG_TAG, status.message);
+                callback.apply(null, new WattsCallbackStatus(false, status.message));
+                return null;
+            }
+
+            SceneManager.getInstance().deleteUserScenes((var1, status1) -> {
+                if(!status1.success) {
+                    Log.e(LOG_TAG, status1.message);
+                    callback.apply(null, new WattsCallbackStatus(false, status1.message));
+                    return null;
+                }
+
+                LightManager.getInstance().deleteLightsForUser((var11, status11) -> {
+                    if(!status11.success) {
+                        Log.e(LOG_TAG, status1.message);
+                        callback.apply(null, new WattsCallbackStatus(false, status11.message));
+                        return null;
+                    }
+
+                    callback.apply(null, new WattsCallbackStatus(true));
+                    return null;
+                });
+
+                return null;
+            });
+
+            return null;
+        });
     }
 
     public static UserManager getInstance() {
